Add agent_policy_id and policy_revision_idx to checkin attributes
Allow the agents to add their currently running policy_id and
revision_idx attributes to the checkin request bodies. These attributes,
if included and different from the agent doc will be used when updating
the agent doc in the pre-poll checkin. If the agent's policy id does not
match the expected policy id from the server a reassign is detected and
a new policy change action will be sent. If the checkin ID is greater
than what was previously recorded or the policy id changes from what was
previously recoreded, then the api keys will be managed.

+++ internal/pkg/api/openapi.gen.go
@@ -332,7 +332,7 @@ type CheckinRequest struct {
 	Message string `json:"message"`
 
 	// PolicyRevisionIdx The revision of the policy that the agent is currently running.
-	PolicyRevisionIdx *int `json:"policy_revision_idx,omitempty"`
+	PolicyRevisionIdx *int64 `json:"policy_revision_idx,omitempty"`
 
 	// PollTimeout An optional timeout value that informs fleet-server of when a client will time out on it's checkin request.
 	// If not specified fleet-server will use the timeout values specified in the config (defaults to 5m polling and a 10m write timeout).
